requirements.txt
streamlit>=1.28.0
PyGithub==1.59.1
pandas>=2.0.0
plotly>=5.17.0
python-dateutil>=2.8.2